attendance_web/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
psycopg[binary]==3.2.6
APScheduler==3.10.4
pandas==2.2.2
openpyxl==3.1.5
python-dotenv==1.0.1
